backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.4
minio==7.2.7
pymupdf==1.24.3
pytesseract==0.3.10
reportlab==4.2.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
pydantic-settings==2.2.1
google-generativeai==0.8.3
groq==0.11.0
Pillow==10.3.0
websockets==12.0
httpx==0.27.0
python-dotenv==1.0.1
aiofiles==23.2.1
